ref(http): factor out common HTTP for reuse
