+++ crates/testing/Cargo.toml
@@ -12,7 +12,7 @@ serde = "1"
 serde_json = "1"
 spin-app = { path = "../app" }
 spin-core = { path = "../core" }
-spin-http = { path = "../http" }
+spin-trigger-http = { path = "../trigger-http" }
 spin-trigger = { path = "../trigger" }
 tokio = { version = "1", features = ["macros", "rt"] }
 tracing-subscriber = "0.3"
