+++ crates/trigger-http/benches/baseline.rs
@@ -6,8 +6,8 @@ use criterion::{black_box, criterion_group, criterion_main, Criterion};
 use futures::future::join_all;
 use http::uri::Scheme;
 use http::Request;
-use spin_http::HttpTrigger;
 use spin_testing::{assert_http_response_success, HttpTestConfig};
+use spin_trigger_http::HttpTrigger;
 use tokio::runtime::Runtime;
 use tokio::task;
