+++ crates/trigger-http/benches/spin-http-benchmark/Cargo.toml
