+++ crates/trigger-http/benches/wagi-benchmark/Cargo.toml
