+++ crates/trigger-http/benches/wagi-benchmark/src/main.rs
