+++ crates/trigger-http/tests/wagi-test/Cargo.toml
