+++ src/commands/deploy.rs
@@ -10,14 +10,14 @@ use hippo_openapi::models::ChannelRevisionSelectionStrategy;
 use rand::Rng;
 use semver::BuildMetadata;
 use sha2::{Digest, Sha256};
-use spin_http::routes::RoutePattern;
-use spin_http::AppInfo;
-use spin_http::WELL_KNOWN_PREFIX;
 use spin_loader::bindle::BindleConnectionInfo;
 use spin_loader::local::config::RawAppManifest;
 use spin_loader::local::{assets, config, parent_dir};
 use spin_manifest::ApplicationTrigger;
 use spin_manifest::{HttpTriggerConfiguration, TriggerConfig};
+use spin_trigger_http::routes::RoutePattern;
+use spin_trigger_http::AppInfo;
+use spin_trigger_http::WELL_KNOWN_PREFIX;
 use tokio::fs;
 use tracing::instrument;
